Tail the logs instead of trying to build a link
AWS make this _super annoying_. `aws logs tail` only supports getting
all of the logs from an entire group (in our case the whole govuk
cluster), so we have to use `grep` on the client side to filter out the
logs we want. Why do you have to make this so bad AWS? Why?