Provide AppMesh Virtual Services as Virtual Node backends
This fixes a misconfiguration of our AWS App Mesh
Virtual Nodes. When Envoy intercepts outgoing
application requests to other services in the mesh,
it checks whether it has a virtual service it can
send the request to. If it does not, it simply forwards
on the request unchanged (no load balancing features
are triggered) if the the egress filter is set to
ALLOW_ALL. If this egress filter is set to DROP_ALL
the intercepted outgoing requests are dropped if they
don't match a registered backend service on the virtual
node.

Our configuration currently doesn't define any backing
virtual services for our virtual nodes (apart from
permitting the virtual services to send requests to
themselves).

Changing our egress filter to DROP_ALL revealed that
all requests we assumed had been load balanced by
Envoy were instead load balanced by Route53 returning
the DNS A record list re-ordered in a round robin fashion.

This changes the service mesh node module such that
one can provide a list of backing virtual services
for a virtual node, so that these virtual services
can be treated as an Upstream Cluster by Envoy.

I'd recommend reading the AWS documentation, but
particularly the Envoy documentation. Envoy is a
client-side (distributed) load balancer/proxy.

In the course of discovering this bug with Richard
Towers (props to Richard for actually spotting this!)
We found that the admin panel and endpoints
for Envoy are quite useful. Use /config_dump and
port 9901 for a good time. Prometheus metrics from
Envoy will also help in future, as will ecs exec.

---

An aside comment on the implementation:
Having the service-mesh node implemented in the app
module is a convenience but we may find we want to
pull this module up and have the services/nodes defined
in the deployment module. We should consider this when
we start adding external services such as DocumentDB,
EC2-hosted databases, Sentry, etc. as virtual services.

For now, we've left the egress filter as ALLOW_ALL
rather than DROP_ALL since it'd be a lot of work
and a confusing debugging experience to block all
outgoing requests. We think we can add virtual services
gradually - this would be worth doing since we'll get
some load balancing properties and observability into
outgoing requests to external services.

Richard also had an interesting idea about using a
black hole IP address for the A record for services,
rather than having CloudMap create DNS records. All we
need is for the applications to resolve the IP address,
and then for Envoy to re-resolve the domain to an IP
from its cache of IPs. The IP address the app gets is
ignored by Envoy if the service is recognised, so this
could work - needs experimentation.