Add path restriction rules to public ALBs
This change enables apps to restrict specified paths to
designated CIDR blocks.

This is useful for apps which have internal APIs such as
Signon's new bearer token API at /api/*.

This also restricts /healthcheck by default for all apps
to our VPC and VPN.

We may want to revisit using count for this, given count
can create unintuitive behaviour. For example, removing or
adding paths or IPs may cause existing rules to be recreated.
There doesn't seem to be a good way to do this, but I've put
in a TODO so we can revisit this.