Add Lambda for rotating Signon bearer tokens stored in SecretsManager
This adds a Ruby script to be run by AWS Lambda. The script will
be triggered by SecretsManager in order to rotate a Signon application
bearer token.

The lambda handler follows the four steps (create, set, test, finish)
defined in the documentation:
https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets-lambda-function-overview.html.

Further work is required to complete the rotation:

* Signon must have an API callable by the Lambda
* In a subsequent commit, the infra for the Lambda will be created
* Signon must have a bootstrap task that can create resources
  such as api_users, applications, for a workspace
* Concourse must cycle the ECS tasks on a schedule to ensure that
  the apps have the current version of the secret
* Signon must delete old, unused secrets, if rotation is enabled

We'll also need to bootstrap application oauth secrets later, with
a similar lambda function.