Log from ECS to Splunk
This commit changes logging for ECS so that logs from both `app` and `envoy` containers move from AWS CloudWatch Logs to Splunk. [Splunk is the recommended tool for storing infrastructure, application and audit logs](https://gds-way.cloudapps.digital/standards/logging.html#how-to-store-and-query-logs).

Splunk has been set-up so that it logs to a new `govuk_replatforming` index, which will need to change as and when we move the new platform into any real environments. As we're using the Splunk Docker driver and aren't changing the logging of individual applications in this commit, we're using the `tag` option in the logging configuration to specify some additional metadata which will be useful for debugging, such as the image name, container name and container ID. The `splunk-sourcetype` option is currently being hard-coded to the value `log`, and the `splunk-format` option is being set to `raw` which enables us to query the values that were specified earlier with the `tag` option from within the Splunk interface.

Finally, the `splunk-url` and `splunk-token` options are set from values stored using AWS Secrets Manager.