Remove .terraform.lock.hcl files from version control
Terraform 0.14 introduced HCL lockfiles. By default generating a
lockfile with `terraform init` will create a lockfile that is only
valid in the platform that you ran terraform init in. There's no
way to override this default behaviour as far as I can see.

For example, on macOS running `terraform init` will autogenerate
a lockfile that will cause `terraform init` to fail on Linux
(the platform we use for CI/CD).

Our current workaround was to run terraform providers lock and specify
all target platforms - we could possibly do this in a pre-commit
hook: https://www.terraform.io/docs/cli/commands/providers/lock.html#specifying-target-platforms.

However, we're not seeing a benefit to committing the lockfil so
we are going to see if we can do without it.