Remove custom_container_services from app module
We don't currently have any apps that need to open
up more than one port, so we can remove the code
from app that enables this.

Some apps may need to open up additional ports in
future, if we are unable to redesign the apps
- this is unlikely.

Some apps provide a /metrics endpoint on an admin
port, such as router.

I would suggest that we extract the service mesh
node logic from app, so that we can pass in multiple
ports to app, and have a mesh service per app port.