Remove service_mesh:bool variable from app
This created a lot of complexity for little gain.
We don't lose anything from Grafana having a
virtual service registered in the mesh, so we don't
need this flag.

I think it is safe to assume that every app will
be part of the service mesh.

However, it may be the case that an app will be
registered as multiple services - i.e. an admin
port vs a public port for a single ECS service.
If this need arises, we should extract the
service mesh node module from the app and have
the clients create them.