Retry failed Signon token rotation requests
AWS Secrets Manager will retry failed lambda invocations.
The retry mechanism is 5 times, once per minute (for a total
of 5 minutes). If the retries fails the invocation will be
retried once per week for 4 weeks. After that the invocation
will be abandoned.

This information is subject to change - we needed to write in
to AWS Support to find this out, since it's not in their docs.

This is a problem for us, as there is a race condition:
the lambda invocations can be exhausted before Signon is up
during the initial bootstrapping process.

This simple patch is to retry the requests to Signon for
a bit (up to the max runtime of 15 mins) in the function
itself. This doesn't solve the race condition, but does
alleviate it.

The proper solution coming in a future commit will be to
create the initial secret versions synchronously during
bootstrapping.