Create initial Signon bearer token versions from Concourse
During the bootstrapping process, the bearer tokens for an
app will be created before the app is deployed (and after
signon is up and ready).

After the initial deploy, the boostrap-secrets task will be
a no-op - secrets will be rotated by the lambdas.

This introduces a new Docker image to run Concourse tasks
from. This will make it easier to test and verify Concourse
tasks - in place of untested Bash scripts.

Still todo: create a pipeline that will build and distribute
the docker images in govuk-infrastructure.

This change will ensure that we can bootstrap app secrets
quickly, without needing to wait a long time for lambdas
to be restarted.