Link to Splunk for logs when updating ECS service
This commit links to Splunk for logs when updating an ECS service. This change is necessary as we're no longer logging to AWS CloudWatch Logs, so the call to `ecs-cli logs` is failing. As we have the `runtimeId` of ECS containers logged to Splunk (via the `container_id` property) we can instead track ECS updates there.