Wait 15minutes for ECS deploy to complete
An ECS service is stable when service.deployments == 0
and runningCount == desiredCount. In other words, when
there are no active deployments and the desired number
of tasks are running.

The ecs wait service-stable command [1][] is used to wait
for a deployment to complete.

However, the polling time for the wait command is not
configurable: it polls every 15s for 40 attempts (10m).

A deploy can take longer than this.
For example the publisher-web deployment this morning
was started at 10:01:37 and completed at 10:16:26.

This change will make the deploy job wait for 20mins
for an ECS deployment to complete. This is a bit
excessive, but seems to be necessary.
If we find this is too long to wait for success, we
can now configure the timeout/attempts per deployment
with this change.

[1]: https://awscli.amazonaws.com/v2/documentation/api/latest/reference/ecs/wait/services-stable.html