Increase content-store ECS task size
Increases Content Store CPU and Memory from
1 vCPU and 2GB RAM to 2 vCPU and 4GB RAM.

Changes to the CPU and Memory params need to
consider the permitted task sizes (e.g. you're
not allowed a 2 CPU units and 1GB RAM).
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#task_size

This change is made since we've seen spikes
of CPU usage when starting content-store's
container, and failing container health checks.
We're interested to see whether the health
check failures will be reduced if we make
content-store's task beefier.