Add authenticating-proxy app
Adds the `authenticating-proxy` to the `govuk-publishing-platform`
which authenticates HTTP requests before it reaches the draft-router by
either authenticating with Signon or validating the token in the HTTP
requests.

Hence, the `authenticating-proxy` replaces `draft-router` as the first
app when in the draft traffic flow.

**NOTE**
authenticating-proxy is named `Content Preview` in Signon in other
environments and this has been kept this way in ECS Signon.

**Testing**

This was tested by deploying this PR in `fred` workspace and browsing
to:
https://draft-origin.fred.test.govuk.digital/ . You should be redirected
to Signon for login. Note that your user should be authorised to use
`Content Preview` app in Signon for Signon to authorise the request and
redirect you to the draft stack.

Ref:
1. [task trello card](https://trello.com/c/vYNGbJSN/493-add-authenticating-proxy-to-the-draft-stack)
2. [authenticating-proxy dockerfile](https://github.com/alphagov/authenticating-proxy/pull/269)