fix grafana deployment
This PR fixes the Grafana deployment that has been broken a while now.
This is because the deployment uses many common modules with
govuk-publishing-platform and some changes were not backported to this
deployment.

This deployment uses only the `public-load-balancer` child module
now and to prevent further breakages, we add a GitHub integration
check for this deployment also.

Ref:
1. [Trello task](https://trello.com/c/twpFac1c/499-fix-grafana-deploy)