Ignore pending rotations when bootstrapping signon tokens
This resolves a race condition which would prevent a
secret value being set for a bearer token during bootstrapping.

If a rotation has started and failed (i.e. will try again tomorrow)
we still want to set a current version during bootstrapping,
regardless of any pending rotations.

Between the bootstrap check for an AWSCURRENT version and
the bootstrap set command, the rotation could complete.
This would not cause an error, since we need either of the
working secrets to be set for the initial run of the pipeline.