Eliminate manual build step for token rotation lambda.
Instead of deploying from a manually-built zip file checked into the
repo, have Terraform build the zip file.

This still isn't an ideal way to deploy a lambda, but it at least avoids
the need to remember to rebuild the zip file manually and check it into
the repo every time.