Use app domains for backend apps
This changes the domains we'll use for backend apps
in signon OAuth app config and when making requests
to signon in the default workspace from the
public domain publishing.service.gov.uk to
the app domain ecs.test.govuk.digital.

This is because we've not yet set up the backend apps
on publishing.service.gov.uk subdomains.