migrate test cdn to govuk.digital domain
We migrate the test CDN from `publishing.service.gov.uk` to the `govuk.digital` domain so as to stop cannibalising the `publishing.service.gov.uk`. This is useful before the spin-up in integration.

We add an option to enable or not the configuration of the `govuk.digital` test CDN, there will be scenarios where we just want to use the `publishing.service.gov.uk` CDN only.