Wait longer for secrets to be set
This change will increase the amount of time
an app waits for secrets to be set before it
attempts to deploy from ~4mins to ~17mins.

We also re-attempt per app in the deploy pipeline

This should stop those errors we saw this morning
when we destroyed and re-ran the deploy pipeline.