Disable locking on plan in CI
I couldn't see a reason for the terraform plan to
take a lock (and cause any run-terraform jobs to
fail). One reason is consistency - you ensure that
nothing modifies the state while you're reading,
but I'd rather have CI plans give weird results
than have run-terraform jobs fail.

This thread had some suggestions that a plan
does write to the statefile as it does a refresh
but that only modifies the local state, not
the remote state:
https://discuss.hashicorp.com/t/why-does-a-plan-lock-the-state-file/3520