Remove splunk logs from update-ecs-service job
We don't have a good way to see the logs for all tasks
on an ECS Service at the moment. One way would be to
feed in GOVUK_APP_NAME to the update-ecs-service job,
for the purpose of linking to logs filtered by the relevant
GOVUK_APP_NAME. This seems a bit pointless though, since
we'll be able to filter easily in Splunk, provided we
know what index to filter by.

Removing this since sometimes we can't retrieve the task
id (and the fix would be to re-attempt this) but don't
actually need it here.