Commit workspace params files to main branch
Props to @smford for this - he implemented this in an earlier
PR.

This includes the pipeline file in the main branch. The benefit
is that we can keep our workspaces up to date without needing
to rebase frequently. This will reduce the amount of time we
need to spend repairing workspaces that fell too far behind.

If we want to work on a separate branch, we can create a new
branch and point our pipeline at that branch briefly.