Make Smokey workspace-aware
This ensures that the smoke tests are run for the correct
workspace.

Currently it is possible for a workspace to run smoke
tests for a different workspace.

Smoke tests currently have a single task definition
family called 'smokey' unlike other task definition
families such as 'frontend-bill' or 'publisher-ecs'.

As you can see, each pipeline will update the smoke
tests with variables such as GOVUK_APP_DOMAIN - a
workspace-aware domain such as bill.test. However,
since the family is not workspace-safe, the smoke
tests will simply run the tests for the workspace
that last update the smokey task definition.

This resolves the described problem by creating a
smokey task family for each workspace.