Wait longer for deploys to complete
This adds an extra 5 minutes to the max wait time
on deploy completions. This should resolve the issues
we've seen where a deploy waiter times out, but the
deploy was successful a minute or so later.

We're learning that it takes up to 25 minutes for
ECS to do a rolling deployment of just one task
with almost no traffic.