Exclude smoke tests not for current environment
This commit excludes smoke tests not meant for the current environment when they're run by Concourse. To facilitate this, any tests not meant for a particular environment should be marked up as `@not<environment>` in Smokey, where `<environment>` is to be replaced with the name of the environment in which they should not be run.

The reason for this change is to allow us to run smoke tests for publishing content, ensuring that they do not run in production.