Autogenerate frontend secret_key_base
This creates the secret for frontend's SECRET_KEY_BASE
environment variable. This replaces the data block that
pulls a secret that we manually inserted into Secrets Manager.

Later, I will add a commit that creates the initial secret value,
at the moment we just export the ARN which is all we'll need
to generate a secret and populate the version in Secrets Manager.