Log Smokey output to AWS
This commit updates the logging for Smokey to be sent to AWS, rather than Splunk. This is necessary so that developers can view the entire output of a particular Smokey job when debugging issues, as logs sent to Splunk are broken up by log line and means navigating multiple search results pages to see the output of a single job.