Exclude traffic to services outside the cluster from App Mesh.
This is to (hopefully) prevent the intermittent TCP failures that we've
been seeing occasionally on connections to MongoDB. It's a
hard-to-reproduce problem so it makes sense to just exclude all this
traffic from App Mesh preemptively, given that there's no benefit to
having it go through Envoy. (We don't plan to use any of the App Mesh
features for this traffic anyway so there's no reason for it to.)

See https://trello.com/c/6yn0REmI/368.