Stop using providers meta-argument for origin module
The origin module does not need us to set the meta-argument
providers [1].

The default behaviour is to inherit the root module's default
providers, i.e. all providers that aren't an alias.

We sometimes use the providers argument when we want to make
an alias provider such as aws.us_east_1 available to a module.
This isn't necessary here, since the origin module doesn't use
aws.us_east_1.

[1]: https://www.terraform.io/docs/language/meta-arguments/module-providers.html