Don't give providers to signon_bearer_token module
The default behaviour is to inherit the default providers,
so by ommiting the providers meta-argument we can do that.

Before upgrading to v0.15.5 I recall needing to use the
providers arg since it's in a for_each block. It looks
like it's not needed any more (it may never have been
needed!).