Allow for different log providers in container definition
This commit changes the `logConfiguration` schema within the `container_definitions` variable of the task definition to allow for different providers. As it stands, the `logConfiguration` is hard-coded to Splunk, however we wish to log to AWS logs so that we can stream output for non-app deployments to Concourse (such as Smokey output, Rake tasks etc).