Make Deploy Event bucket environment specific
The deploy event bucket is used for triggering deploys.
It enables us to put an event object into the bucket
whenever we want to trigger a deploy. Who needs Kafka
when you have S3.

However the bucket must be environment-specific. I forgot
that S3 bucket names are global (not scoped to an account).

This fixes the problem by putting the govuk_environment into
the name of the event bucket.