use IAM roles with registry-image Concourse resource
Currently, the Concourse `registry-image` resource is using long life
credentials of an IAM user that has been created to read from the GOV.UK
production ECR. We want to switch to IAM role and get the
credentials from the role attached to the EC2 instance that
runs the Concourse pipeline.

This will brings us the 2 main benefits:
1. short live credentials are used and less likely to be useful if
compromised.
2. there is no need to create credentials in the AWS console and add it
to the Concourse team secrets for every new team/GOV.UK environment that
we create.

Unfortunately, the Concourse `registry-image` resource does support
(i) getting AWS credentials from EC2 metadata and (ii) following a IAM
roles chain: this is required since the EC2 instances have the
credentials of the Concourse worker role which needs to be used to
assume the Concourse deployer role. The deployer role is then used to
assume a role in the GOV.UK production ECR to pull images.

A fork for the Concourse `registry-image` resource is available
[here](https://github.com/alphagov/registry-image-resource)

Signed-off-by: Frederic Francois <frederic.francois@digital.cabinet-office.gov.uk>