Provide content-schemas to publishing-api at startup
When Publishing API ECS tasks startup, a transient volume will be
attached to the task (shared between tasks), and an init container
will be run before the app container is started.

The init container will copy the contents of the content-schemas
S3 bucket in the workspace to the path /govuk-content-schemas in
the volume.

The content schemas can be updated by redeploying publishing-api -
this is managed by Concourse (see prev commit).