Fix dynamic vars for tasks
Provides vars to task step files. When vars such as
((my-var)) are defined in a task step file i.e. my-task.yaml,
then callers of that step file need to specify the value of the
var in the `vars` param for that task.

See the vars field here for more detail:
https://concourse-ci.org/jobs.html#schema.step.task-step.file

If this field is not provided you'll see an error that looks like:
failed to interpolate task config: undefined vars: my-var.