Add app dependencies to deploy pipeline
This change uses the bootstrap-trigger and Concourse passed field

to ensure that apps are brought up in the correct order.

For example, for publisher's smoke test to pass, frontend
must be deployed first. Likewise, for publishing-api to
be deployed, signon must be running.

This change makes app dependencies explicit, which is also
useful.

The change only has an impact during bootstrapping. After
creating the infrastructure apps will be deployed independently
provided that their dependencies have been brought up
previously.