Add Concourse pipeline for terraforming EKS
This commit creates a pipeline for terraforming EKS, consisting of two jobs `terraform-eks-cluster` and `terraform-eks-cluster-addons`, which terraforms the EKS cluster and the addons respectively. This is necessary for the reasons as detailed in commit c228db6.

We also change the way we assume the `govuk-concourse-deployer` role.  Instead of configuring the role once per Terraform provider per top-level module (`deployments/*`), we do the assume-role as part of the terraform-apply.sh script (by running aws sts assume-role). This results in less overall complexity. (Config for the k8s TF provider to assume a role is different from the equivalent config for the AWS provider.)

@fredericfran-gds pointed out that we could pass `--role-arn` to the `aws` command from the k8s TF provider config. Our approach however means we have one place and one way of assuming the role, rather than two different ways and multiple places (one per provider per top-level module).

We also add the `govuk-concourse-deployer` role to the `aws-auth` configmap so that Concourse can apply changes to the Kubernetes cluster via the Kubernetes Terraform provider. This is only strictly necessary when Concourse is modifying an existing cluster _not_ created by Concourse (i.e. not created by the `govuk-concourse-deployer` role), but we're including this for robustness and to make the required permissions explicit.

We're running Terraform in automation mode per the best practices described in https://learn.hashicorp.com/tutorials/terraform/automate-terraform

Co-authored-by: Chris Banks <chris.banks@digital.cabinet-office.gov.uk>