Add CI Terraform Plan for EKS
Since we switch to EKS from ECS, this PR adds the terraform plans for
the EKS modules and removes the ECS ones.

**Testing**

Tested by setting the CI pipeline to the one defined here and
checking that the 2 terraform plans are done and reported to
GitHub

**Notes**
We use plan with `-lock=false` and `-refresh=false` as done
previously in ECS with justifications in #297 and #306 respectively.

Ref:
1. [EKS cluster](#372)
2. [EKS cluster addons](#377)
3. [Concourse EKS pipeline](#378)
4. [add CI user to k8s](#391)