Remove the Concourse deployer IAM role from the aws-auth ConfigMap.
Including this was necessary in order for Concourse to manage objects in
a cluster which Concourse didn't originally create, but it introduces a
risk of locking ourselves out of the cluster entirely if we break the
aws-auth ConfigMap.

If we ever did need Concourse to "adopt" a cluster which it didn't
create, we could still do this by adding its role to the ConfigMap at
that time. Unless/until that's needed, it's safer to omit it so that we
can always access the cluster via the cluster creator IAM role (which
from this point on will be the govuk-concourse-deployer role).