add readonly users to aws-auth config map
This commit adds `readonly` users (as defined in
[govuk-aws-data](https://github.com/alphagov/govuk-aws-data)) are added
to the cluster.

These new users are mapped to the k8s group `readonly` with the default
[view](https://kubernetes.io/docs/reference/access-authn-authz/rbac/)
clusterrole is associated to that group.

Since Concourse CI in the `test` environment needs a user to verify
commits via terraform plans, we added a new readonly user
`govuk-ci-concourse` to the cluster in the `test` environment.

We add 2 additional privileges to the `readonly` k8s group: read
clusterroles and clusterrolebindings since this necessary for
`govuk-ci-concourse` to be able to plan the `cluster-services`
deployment.

Ref:
1. [trello card](https://trello.com/c/ONhtNSTg/406-iam-eks-configure-fulladmin-and-readonly-access-levels)