Rename `eks` and `eks-stage2` Terraform projects
This commit renames the `eks` and `eks-stage2` Terraform projects to `cluster-infrastructure` and `cluster-services` respectively. This has been defined in a [recent ADR on splitting the Terraform state](https://github.com/alphagov/govuk-infrastructure/blob/main/docs/architecture/decisions/0003-split-terraform-state-into-separate-aws-cluster-and-kubernetes-resource-phases.md#decision) for the EKS and Kubernetes resources and gives clearer meaning to the intentions of the projects:

- `cluster-infrastructure` is concerned only with setting-up the EKS cluster and all associated AWS resources (such as the worker groups and auto-scaling groups).
- `cluster-services` is concerned only with setting up the Kubernetes resources and configuration, including the `aws-auth` ConfigMap.

As part of this commit, the Terraform state for the existing `eks` and `eks-stage2` projects has been duplicated within the state bucket and renamed appropriately for the renamed projects; once we transition over to using the renamed projects, the old Terraform state files can be deleted.