Give the EKS cluster its own subnets and associated resources.
The VPC itself and the Internet gateway are still defined in
alphagov/govuk-aws for now, since the current plan is to use the
existing VPC and move/import it here after the migration to Kubernetes.

This follows the guidance in
https://docs.aws.amazon.com/eks/latest/userguide/network_reqs.html.

In each of three AZs, we have a public /28 for the control plane (i.e.
the kube-apiserver VIP) as recommended, a public /24 for Internet-facing
load balancers and a private /22 for nodes and pods.

Separating the node and pod subnets is possible but involves some [hairy
configuration of Amazon's CNI plugin][cni-custom-network], so it doesn't
seem worth it without a specific need.

Fetch the VPC ID from infra_vpc instead of infra_networking, since this
lets us drop infra_networking remote state while still being able to
retrieve the Internet gateway ID.

[cni-custom-network]: https://docs.aws.amazon.com/eks/latest/userguide/cni-custom-network.html