Include cluster name in names of all managed AWS resources.
This allows for testing of breaking changes using workspaces.

Deliberately not including terraform.workspace anywhere in the config,
so as to avoid a repeat of the toil and complexity of the ECS TF codebase.
I wanted to have var.cluster_name default to the workspace name if
workspace!=default, but sadly this isn't possible in Terraform. (srsly?)
I also didn't want to introduce local.cluster_name because that's just
needlessly confusing.

So the way to use workspaces (on the hopefully-rare occasions that
they're needed) is:

    tf workspace new $USER

    tf apply \
        -var-file=../variables/test/common.tfvars \
        -var cluster_name=$USER

    tf destroy \
        -var-file=../variables/test/common.tfvars \
        -var cluster_name=$USER

Once the VPC stuff is merged (and the cluster has its own subnets),
you'll also need to define a local vars file with a unique set of
subnets so as not to overlap with the main test cluster.